+++ pkg/output/util.go
@@ -1,0 +1,49 @@
+// Copyright 2024 The KitOps Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package output
+
+import (
+	"fmt"
+	"math"
+)
+
+func FormatBytes(i int64) string {
+	if i == 0 {
+		return "0 B  "
+	}
+
+	if i < 1024 {
+		// Catch bytes to avoid printing fractional amounts of bytes e.g. 123.0 bytes
+		return fmt.Sprintf("%d B  ", i)
+	}
+
+	suffixes := []string{"KiB", "MiB", "GiB", "TiB"}
+	unit := float64(1024)
+
+	size := float64(i) / unit
+	for _, suffix := range suffixes {
+		if size < unit {
+			// Round down to the nearest tenth of a unit to avoid e.g. 1MiB - 1B = 1024KiB
+			niceSize := math.Floor(size*10) / 10
+			return fmt.Sprintf("%.1f %s", niceSize, suffix)
+		}
+		size = size / unit
+	}
+
+	// Fall back to printing whatever's left as PiB
+	return fmt.Sprintf("%.1f %s", size, "PiB")
+}
